Mulighet for id-lenker
